Register IClientItemExtensions / IClientFluidTypeExtensions properly
