+++ src/main/java/com/lance5057/butchercraft/armor/BootsItem.java
@@ -1,29 +1,11 @@
 package com.lance5057.butchercraft.armor;
 
-import com.lance5057.butchercraft.armor.models.BootsModel;
-
-import net.minecraft.client.Minecraft;
-import net.minecraft.client.model.HumanoidModel;
-import net.minecraft.client.model.geom.EntityModelSet;
-import net.minecraft.client.model.geom.ModelPart;
 import net.minecraft.core.Holder;
-import net.minecraft.world.entity.EquipmentSlot;
-import net.minecraft.world.entity.LivingEntity;
 import net.minecraft.world.item.ArmorItem;
 import net.minecraft.world.item.ArmorMaterial;
-import net.minecraft.world.item.ItemStack;
-import net.neoforged.neoforge.client.extensions.common.IClientItemExtensions;
 
-public class BootsItem extends ArmorItem implements IClientItemExtensions {
+public class BootsItem extends ArmorItem {
 	public BootsItem(Holder<ArmorMaterial> pMaterial, Properties pProperties) {
 		super(pMaterial, Type.BOOTS, pProperties);
 	}
-
-	@Override
-	public HumanoidModel<?> getHumanoidArmorModel(LivingEntity livingEntity, ItemStack itemStack, EquipmentSlot equipmentSlot, HumanoidModel<?> original) {
-		EntityModelSet models = Minecraft.getInstance().getEntityModels();
-		ModelPart root = models.bakeLayer(BootsModel.LAYER_LOCATION);
-		return new BootsModel(root);
-	}
-
 }
