+++ src/main/java/com/lance5057/butchercraft/armor/BunnyTailItem.java
@@ -1,29 +1,11 @@
 package com.lance5057.butchercraft.armor;
 
-import com.lance5057.butchercraft.armor.models.BunnyTailModel;
-
-import net.minecraft.client.Minecraft;
-import net.minecraft.client.model.HumanoidModel;
-import net.minecraft.client.model.geom.EntityModelSet;
-import net.minecraft.client.model.geom.ModelPart;
 import net.minecraft.core.Holder;
-import net.minecraft.world.entity.EquipmentSlot;
-import net.minecraft.world.entity.LivingEntity;
 import net.minecraft.world.item.ArmorItem;
 import net.minecraft.world.item.ArmorMaterial;
-import net.minecraft.world.item.ItemStack;
-import net.neoforged.neoforge.client.extensions.common.IClientItemExtensions;
 
-public class BunnyTailItem extends ArmorItem implements IClientItemExtensions {
+public class BunnyTailItem extends ArmorItem {
 	public BunnyTailItem(Holder<ArmorMaterial> pMaterial, Properties pProperties) {
 		super(pMaterial, Type.LEGGINGS, pProperties);
 	}
-
-	@Override
-	public HumanoidModel<?> getHumanoidArmorModel(LivingEntity livingEntity, ItemStack itemStack, EquipmentSlot equipmentSlot, HumanoidModel<?> original) {
-		EntityModelSet models = Minecraft.getInstance().getEntityModels();
-		ModelPart root = models.bakeLayer(BunnyTailModel.LAYER_LOCATION);
-		return new BunnyTailModel(root);
-	}
-
 }
