+++ src/main/java/com/lance5057/butchercraft/items/SoapItem.java
@@ -1,22 +1,13 @@
 package com.lance5057.butchercraft.items;
 
-import org.jetbrains.annotations.Nullable;
-
 import com.lance5057.butchercraft.ButchercraftMobEffects;
-import com.lance5057.butchercraft.client.ArmPoses;
-import com.lance5057.butchercraft.client.rendering.RenderUtil;
-import com.mojang.blaze3d.vertex.PoseStack;
 
-import net.minecraft.client.model.HumanoidModel;
-import net.minecraft.client.player.LocalPlayer;
 import net.minecraft.core.particles.ParticleTypes;
 import net.minecraft.server.level.ServerLevel;
 import net.minecraft.server.level.ServerPlayer;
-import net.minecraft.util.Mth;
 import net.minecraft.world.InteractionHand;
 import net.minecraft.world.InteractionResultHolder;
 import net.minecraft.world.entity.Entity;
-import net.minecraft.world.entity.HumanoidArm;
 import net.minecraft.world.entity.LivingEntity;
 import net.minecraft.world.entity.player.Inventory;
 import net.minecraft.world.entity.player.Player;
@@ -25,9 +16,8 @@ import net.minecraft.world.item.ItemStack;
 import net.minecraft.world.item.ItemUtils;
 import net.minecraft.world.item.UseAnim;
 import net.minecraft.world.level.Level;
-import net.neoforged.neoforge.client.extensions.common.IClientItemExtensions;
 
-public class SoapItem extends Item implements IClientItemExtensions {
+public class SoapItem extends Item {
 
 	public SoapItem(Properties pProperties) {
 		super(pProperties);
@@ -82,32 +72,5 @@ public class SoapItem extends Item implements IClientItemExtensions {
 				}
 			}
 		}
-	}
-
-	@Override
-	@Nullable
-	public HumanoidModel.ArmPose getArmPose(LivingEntity entityLiving, InteractionHand hand, ItemStack itemStack) {
-		if (!itemStack.isEmpty()) {
-			if (entityLiving.getUsedItemHand() == hand && entityLiving.getUseItemRemainingTicks() > 0) {
-				return ArmPoses.SOAP_ENUM_PROXY.getValue();
-			}
-		}
-		return HumanoidModel.ArmPose.ITEM;
-	}
-
-	@Override
-	public boolean applyForgeHandTransform(PoseStack poseStack, LocalPlayer player, HumanoidArm arm, ItemStack itemInHand, float partialTick, float equipProcess, float swingProcess) {
-		float f = player.getUseItemRemainingTicks() - partialTick + 1.0F;
-		int i = arm == HumanoidArm.RIGHT ? 1 : -1;
-		float i2 = arm == HumanoidArm.RIGHT ? -0.5f : 0.5f;
-		poseStack.translate(i * 0.56F, -0.52F, -0.72F);
-		float f2 = Mth.abs(Mth.cos(f / 8.0F * (float) Math.PI) * 0.5F);
-		if (player.getUseItem() == itemInHand && player.isUsingItem()) {
-
-			poseStack.mulPose(RenderUtil.createQuaternion(0, 90, 0, true));
-			poseStack.translate(0.0, f2, i2);
-
-		}
-		return true;
 	}
 }
